Further intelligence added to banks structure.
New instruments found in banks will be added if the bank isn't
already full - provided they have either the yoshi or zyn extension.
A location prefix will be added to any that don't have one.

If a new instrument has the same prefix and name as an
existing one but a different extension, then if it is a yoshi
type the bank data will be changed to that.

If a new instrument has the same prefix but different name as an
existing one, if the bank is not full a space will be found for
it starting from 160 and counting back. If it is in a writable
area the file will be renamed to match the new location, otherwise
the existing location will be retained, but the displayed one will
be the new location.

An Instrument consisting of just a number and extension, will have
that number turned into a name, and a new location found for it.