Further improvements to bank management
If a file in a bank is seen with the same number prefix as an existing one
then if the *internal* names are the same it is assumed that one or other
simply had a slightly different filename, and therefore is a duplicate to
be ignored.

If the internal names are different it is regarded as a similar but changed
version, in which case a new prefix will be found for it.

If this is in a writeable area the file will actually have it's prefix
changed, but if it is not writeable, the displayed prefix will still be
changed, but linking back to the actual file.

Most of this work is done in the addtobank() function, and this has been
considerably re-written with additional comments to clarify the process.