Powf: rationalise usages of decibel factors
- use a common front-end function to get amplitude factors from dB parameters
- dB parameter can be mapped to a scale (using the template param). eg -60dB scale
- implement all dB factors with the optimised power<base>() function
- remove optional use of the exp10f() function

Rationale:
We want reproducible values, even under strong optimisation.
Up to now, some optimisers were clever enough, to replace powf(base, exponent)
with exp(ln(base)*exponent). And if that happens, the generated samples will
be slightly different, typically off by a tiny amount ~ -124dB

The exp10f() function is not (yet) standard. If it would be used, again
the values would be slightly different (actually a bit more correct).
The delta in this case is also around -124dB

However, unfortunately the CMake-Macro check_function_exists is known
to be unreliable, and frequently fails to detect such ubiquitous functions
like log, exp, exp10f (because these are typically mapped with macros
to a low-level hardware based implementation). As a consequence, in
many cases we analysed, exp10f was not detected and the powf()-based
implementation used instead.

With this changeset, these possible numeric divergences are removed