no-gui: ensure FLTK not activated when Option "ShowGui" is false
Yoshimi can be built with GUI_FLTK not defined, in which case
all GUI actions are left out. But when compiled with Gui but
started with --no-gui, the FLTK-Toolkit was still activated
and attempted to connect to the display -- yet all windows
remained hidden.

This can be problematic when using Yoshimi in a commandline-only
environment, e.g. when running under Docker.