Fix compile with Clang: initialise static var out-of-class
The "constexpr" was used here solely to allow in-class initialisation,
while in fact we do not need a compiletime expression, since we are
initialising a static variable at program start anyway.

Clang complains that the initialiser can not be constexpr.
GCC is more tolerant here (which is a non-standard extension)

log, exp, and similar math functions can not be constexpr,
at least when taking the standard literally.