[padthread] switch wavetable computation to use the new storage
- use the new storage structure as drop-in replacement
- also remove the existing storage definition an all code for manual memory cleanup

Note: the new PADTable is defined in a way that it is always usable and valid.
Even if no spectrum has been rendered yet, the samples are initialised to 0.0
and the basefrequency is set to a fallback value of 440 Hz.

Thus it is no longer possible that some wavetable is NULL.
PAD_MAX_SAMPLES is no longer necessary, since all Wavetables now
reside within a single chunk of contiguous heap memory.