[padthread] turn the spectrum holder (FFTFreqs) into a RAII type
...which enables automatic memory management / clean-up.
Instead of using the helper functions from the FFTwrapper,
this struct is now a class, which ensures the data is always
initialised and is automatically cleaned-up whenever the FFTFreq
object goes out of scope. Also changed various pointer arguments
into references (which can not be NULL).

Motivation for this changes was the ability to "simply walk away"
from the PADnoteParameters::applyparameters() whenever we want
to abort that backgound thread.