[padthread] rationalise usage of frequency index
This is a nasty and dangerous change; however the existing code
was inconsistent with respect of labelling the Partials,
which should be corrected.

I found out the following

 * SLOT-0 of the FFT-Spectrum is the DC-Offset
 * OscilGen uses the FFT such that each Slot corresponds to a Partial;
 * For this reason, it always discards SLOT-0 and places instead
   the coefficients for the fundamental at index position(s) 0

This Changeset now changes some internals of PADSynth to align with this convention.
Index = 0 is the fundamental, Index = 1 the 1st overtone.

The comment at the Function getNhr() "Get the harmonic(overtone) position"
was there already in the first Import by Cal (2010-10-16) and is a typical
careless and sloppy misleading comment. This function does not calculate
/overtones/, it calcluates a factor for each /partial/ relative to a base frequency.
Thus, for the 0th partial (the fundamental), the un-distorted case yields 1.0
as a frequency factor.

This observation motivates why I hereby propose to rename that function for clarification.