[padthread] rearrange and simplify code to use C++ containers
Various refactorings and equivalence transformations to make the code
shorter and improve legibility and safety. Get rid of some further raw arrays.
Since the spectrum and profile arrays were initialised to zero anyway, we can
use std::vector instead, which allows to pass the size implicitly to helper
functions, instead of using an extra size argument or even worse, use some
common global parameter.

Moreover, reorder and split the calculation of the harmonic profile and bandwidth.
This allows to push some variables down, simplify signatures and avoid having to
build the profile when it is not needed (it is needed only for the bandwidth mode).

Rationalise usage of signed / unsigned values and use size_t for indices whenever possible.

Note: mostly, this changeset is (hopefully) an equivalence transformation.
With one exception: I changed the normalisation calculations from float to double,
to avoid number precision problems, which tend to lead to non-portable behaviour
and could be amplified here through the following inverse fourier transform.