[padthread] Draft the state management and scheduling front-end
This is a design draft, based on a future/promise to pass on the build result.

The intention is that PADnoteParameters holds one instance of
FutureBuild<PADTables>

This will be the sole access point for all other code to deal with this
background task running and building functionality. Moreover, in the
BuildScheduler.cpp there will be an actual task runner implementation,
which can be customised (and evoled) independently from the front-end
and the client code. This could use a thread pool later.

===========================================
Including a draft how rescheduling can work

If the computation result was /not/ marked as successful,
we can without problem invoke the /builder function/ recursively
to repackage the same functor and the existing promise anew.

This works without endless recursion since the recursive call does not
happen from the builder function, but from within the functor created
by this builder function.

And, there is another precondition: the BuildOperation itself must ensure
to abort a computation based on some external state, not due to an error
within the computation itself. And moreover the rescheduling must somehow
"heal" that external state.