[FFT-storage] Refactor storage: disentangle calculation of adaptive harmonics
The original code attempted to "re-use" the same functions both
for processing of the real spectrum, and for processing the displayed
spectral lines in the GUI. Yet for some unclear reason, this re-use
was done only for the second part of the calculation, while the first
part worked redundantly on the spectral data of the OscilGen; moreover
both parts used a different index convention.

As after rearranging and realigning, it turned out that the whole
computation is completely redundant in the sin/cos components of the spectrum.

By passing the actual accessor (either to the sine part, or the cosine
part, or an array with a line spectrum) as functor, it is possible
to collapse and simplify that code without loosing flexibility