[FFT-storage] Refactor storage: link Oscil size parameter to FFT size
The original code (which in this part was largely unchanged since
the initial import by Cal in 2010) used to rely on the "magical convention"
that OscilGen just happens to be used within a Synth with a specific OscilSize.

Such magic conventions are hard to maintain and can be broken accidentally.

Thus it is better to link those size parameters to the actual data size.
Moreover, in-depth analysis showed that OscilParameters::basefuncSpectrum
was defined to MAX_OSCIL_SIZE, but only ever used synth->halfoscilsize slots.
So we can now coherently use the table size from the FFT wrapper and we
can check consistency with assertions. This is also a step towards the
goal of getting rid of all those pervasive "back-links" into the SynthEngine,
which seriously hamper further attempts towards concurrency. In most cases,
seemingly the synth-pointer is just abused to access system global parameters.