[padthread] Fix "synchronous Apply" implementation -- for Scripts and Test-Suite
As it turned out, my first attempt was way too naive.

- we can not directly call buildNewWavetable(),
  since the SynthEngine now relies on the new PADtables
  to be published within the FutureBuild instance.

- we can not just calculate a wavetable and push it into that
  FutureBuild, since other builds might run concurrently

- we have to care for the situation that a build is finished,
  but not yet picked up by the SynthEngine, which would delay
  the start of any further build until after the swap. And
  this situation especially arises when running the Test-Suite,
  causing the changed wavetable to be not ready before the test
  is launched.

The solution of all those problems is to have a dedicated implementation
for "synchronous Apply", which proceeds with several fences to ensure
the state transition happens as assumed

- first we have pick up a build possibly still going on when calling APPLY
- if such a build exists, we have to wait for it to finish, then discard its result
- only then a new build can be triggered
- moreover we have to wait for the new build really to start in the background
- at that point, we can again extract the future for that build
- and wait for it to become ready
- this time we can publish the result back, to be picked up by the SynthEngine

*Note* a twist regarding automatic testing:
This "padthread" refactoring requires the Synth thread to pick up a new wavetable,
since the build now happens concurrently, and we can not just stash it into the
data structures used by the Synth (as the old solution could do while the whole
component was muted). Unfortunately this means that the Synth thread has now
to carry some additional synchronisation overhead. This is neglectable in
real world usage, since new instruments will be loaded only occasionally.
However, the automated test suite builds a new wavetable prior to each
test run, and thus would now capture that additional timing overhead.
For this reason, we explicitly store the new wavetable into the
target in SyntEngine::setReproducibleState(seed), which is
called by TestInvoker before start of timing measurement.

*Note-2*
At this point we also need to force recomputation of Spectrum from scratch
...otherwise there could be a race when changing OscilGen parameters in the Test-CLI-Scritpt:
If using background build, an ongoing build could "reap" the dirty state,
and consequently the OscilGen will not be recomputed at Test start,
leading to a different sequence of random numbers.