[padthread] Run "synchronous Apply" from the Cmd-Dispatch (background) thread
It used to be that way before all the reworking of PADSynth wavetable building.
Seemingly also better, since we don't block the Synth-Thread then, only the
Command dispatch and indirectly also the CLI.

However, the non-blocking wavetable builds are still triggered directly from
the Synth-Thread (commandSendReal -> commandPad() ), since there is no reason
to go through the background Cmd thread (sortResultsThread). For that reason,
we now distinguish in InterChange::commandPad() if the requested "Apply" is
really blocking, and forward to indirectTransfers only in that case.

Remark: cleaning up an "evil hack"
The ENUM PART::control::padsynthParameters "magically happened" to have
the same integer value 104 as PADSYNTH::control::applyChanges

It cost me literally hours to find out thus how the dispatch of "Apply"
to the background thread works.