[padthread] handle consecutive loading into the same Part
Scenario
 - no note playing
 - load one instrument -> causes background build
 - load further instruments
 - then play first note

The existing solution handles this situation correct, but not "clean";
because the fist background build will prevent triggering further builds,
until the first note has picked up the result. Instead, just the "dirty"
state will be activated. Now, depending on the timings, this will either
lead to aborting the build, or when the first build is already done,
it will cause re-trigger after loading the first build's results.
Thus, the 1st note will sound the 1st instrument, and the follow-up builds
will kick in.

With this change, loading an instrument instead clears out pending results
of an previous build; unfortunately this requires a blocking wait. In most
cases, this is harmless (and will block only the Cmd-Dispatch). However,
when loading via MIDI program change, the MIDI-Thread might be blocked.