[padthread] Fix "1st note problem" - mark PADnoteParams as updated after build
The PADtable resides within the Parameters object, not within the note itself.

The refactorings on this »padthrad« branch significantly changed semantics here,
we possibly get a later asynchronous update here whenever the Background build succeds.
Thus, initially a PADnoteParams is set up and marked "up to date", but when the
new wavetable was computed, the number of tables or other significant parameters
might have been changed, so the Parametres need to be marked "dirty".

At the start of the next Synth buffer cycle, a new wavetable will be picked up
and after that, all of the note parameters need to be re-checked an recomputed.


Before these refactorings, once a note was constructed, it was basically consistent,
because the PADtables where build synchronously and blocking at that point.
The bug (to miss this delayed state change) caused sometimes the first note
to be garbled, and only with the next note the computenoteparameters() fixed that.