Fix incomplete reset/clean-up of effects
The EffectMgr::cleanup() -- which is called from SynthEngine::shutUp()
and various further state changing functions, failed to also clean up
the effect output buffer -- and this caused previous state to leak
into the Reverb's feedback loop.

Thus "cleanup" was not complete, and seemingly this was amended
by explicitly clearing the buffer in EffectMgr::changeeffect(),
instead of properly doing that in cleanup() -- which is called
from changeeffect() just two lines above.

To be clear: this sloppiness can be traced back to the code
initially imported when Yoshimi was forked in 2010