Allow for clean reset of InterpolatedValue within effects
This resolves further problems with non-reproducible values in Reverb
(and presumably other effects likewise). Now Effect::cleanup() causes
all InterpolatedValues to be "pushed forward to the target value" immediately,
and thus to start out clean with the currently intended target.

Moreover, under some circumstances InterpolatedValue would not stop interpolation,
but rather continue to interpolate between the same old / new value.
This could happen when interpolationLengh is not divisible by the samples increment;
thus the attempt to make the both advanceValue()-variants mostly symmetrical,
so to have a dedicated end condition check in both variants.

While on that refactoring, a memory leak in EQ was discovered and fixed;
the two AnalogFilter instances in each band are now managed by unique_ptr automatically.