Implement fallback BPM.
It is available from the master window, and provides a way to set the
global BPM when the MIDI driver does not provide it, for example
because there is no ALSA device available to send MIDI clock
messages. It becomes active whenever this information is missing.

For both Jack and LV2, it is trivial to add support for this. For ALSA
it's more complicated, because we need to detect that a certain window
has passed before we can conclude that MIDI clock information is
missing.

As part of this, I rewrote the ALSA MIDI loop to be more efficient by
avoiding constant looping and polling. Instead, use the system call
`poll()`, where we can both sleep and set a timeout at the same
time. This has the advantage of reacting instantly to events, but also
provides a long time where the CPU can sleep if there are no
events. I've set the timeout to 500ms so that quitting will still be
relatively quick.

Signed-off-by: Kristian Amlie <kristian@amlie.name>