[padthread] Implement Legacy Mode with blocking PADSynth-apply
This provides the option "muted", which closely mimics the behaviour
of previous Yoshimi versions: Changes need to be applied manually,
in which case the current PADSynth part is muted, while the wavetable
computation is then carried out right within the current thread, blockingly.
Thus, the PADnoteParameters::buildNewWavetable() needs to be invoked from
the low-priority (event handling) thread, which means no further notes
can be dispatched, but existing notes will continue to sound the other
non-PAD parts until the new wavetable is ready.

Note: in fact this mode avoids all background thread scheduling and
synchronisation primitives and thus can be used on old machines with
limited resources; actually swapping the new wavetable into the Synth
without synchronisation is not correct, just known to work on average.