Undo envelope points now working correctly, but had to create a special case for this.
Normally the undo code looks for and stores a controls previous state,
but when you are creating a point there is no previous state to store.

The answer was to have a routine that made an undo entry using the values
that were going to be set. Then in the undo/redo code, test for data.insert
being envelopePoints, and if so, skip the read, and copy across the existing
data. Finally, in the actual envelope code we ensure that only a new entry
is added to the undo list. If it gets an undo loopback delete the insertion,
but if it gets a redo loopback reapply the entry returned from the undo code.