[padthread] use equal power mix for the crossfade
Rationale: typically the old and the new waveform have very low correlation,
and the phases are mostly different, even while the spectrum remains consistent.
This is the typical situation where a linear cross fade yields a perceived lower volume
during the fade; we thus need to keep the sum of power constant, i.e. old² + new² = 1

Now unfortunately this requires to use a square root to find the complementary
fade-out curve; in theory, the square root computation could be approximated,
but this is black magic and largely dependent on the underlying hardware.
Thus a better solution is to use segment wise linera fades and to perform
the more expensive guiding curve calculation only once per block.

Note: the X-Fade, and even single segments can span the boundraries of a
single invocation (to calculate cntSmp), since cntSmp <= bufferSize