[Legato] rework and simplify Legato and Portamento
The original code base used a rather tricky scheme for Legato notes:
 - there was always a pair of notes activated
 - one of these notes was kept dormant
 - when starting the next legato note
   - the current note was "teleported" over to the dormant note
   - then the current note was re-initialised with the new note values
   - and the other note conducted a seamless fade-down

This solution was needlessly "clever" and thus caused a host of problems,
like memory leaks and inconsistencies all over the place.

And now (for the "padthread" feature) we're facing the challenge to
integrate the Wavetable cross-fade with possibly several Legato notes
spawned concurrently during the an ongoing fade.

Thus we decided to rework the handling of Legato notes...

 - now Legato notes are allocated to a pos like regular notes
 - the next Legato note is created with the copy constructor
 - and then both are just instructed to fade up / down
 - the old Legato note, when done, is discarded like a regular note

 - only Legato-Portamento notes use a single note instance now.

Alongside with this refactoring the differences between non-Kit mode
and kit-mode in NoteOn have been levelled.