[Legato] adapt state handling within note instances
As a second step, change the logic within the note objects
to fit in with the new Legato scheme.

- all state copying is done by the copy constructor
- the handling of Legato-Portamento was split into a dedicated function
- NOTE_KEEPALIVE is now obsolete, due to simplified Legato handling in Part::NoteOn
- NOTE_LEGATOFADEOUT as explicit transitory state, similar to an Envelope after release

Remark-1: instead of defeating NoteOff by trickery (assigning a new midi-note to "posb",
we now explicitly skip the releasekey() handling when noteState == NOTE_LEGATOFADE...

Remark-2: as it turned out, amplitude values can turn over to negative numbers; this
is deliberately allowed as an effect (e.g. AmpLFO). Thus aboveAmplitudeThreshold()
in SynthHelper was not correct and could produce division by zero and false triggers.