[Legato] sharpen trigger condition for history of mono-Notes
When using the non-polyphonic modes (Mono and Legato),
prssing a key overrides any previous keys, even if they are sustained.
However, when releasing the last key while still holding these previous keys,
we can expect the previous key to become "uncovered" and played again.

While this behaviour is not changed, with the reworked Legato handling it
is now possible to hold and sustain a chord with polyphonic notes and then
to switch over to Legato mode seamlessly. Thus the condition for "uncovering"
previous notes needs to be more restrictive