[padthread] Test support augmented
- allow 500 repetions for extended tests
  turns out this is relevant to minimise the impact of the initial setup

- do not restore or persist configuration established for an automated test case

Augment TestInvoker to place generic events on a test timeline

Up to now, the TestInovoker generated a hard coded noteOn and noteOff event,
and calculated the required number of Synth buffer-cycles explicitly.
But to be able to test legato notes and wavetable swapping (and cross fades),
we need the ability to place note spans into a common timeline.

Based on this new framework, it is now possible to cover Legato and XFades
- use the new event based TestSeq framework to launch overlapping notes
- ability to capture a secondary wavetable and swap it during test execution