Clang complile fix
this usage is a boundary case ...
Clang adheres more to the letter of the spec here, which boils down to
"a complex mathematical function can not be constexpr, since it may set
an error status flag by side-effect" -- while GCC is closer to the typical
programmer's intention, insofar GCC detects that the expression as a whole
is completely determined at compile time and just replaces it by result value.

But actually we don't need a compile time constant here,
the intention was just to have all the math inline and close at hand,
so rather just let the optimiser figure out it's constant and be happy...

PS: clang detected a serious programming bug in ADnote.cpp -- so kudos to Clang!