Bugfix: Legato-Portamento in Kit-Mode
During the »padthread« effort it became necessary to rearrange
the handling of Legato notes.

see: 4fa7686c46f7ca

As part of this refactoring, portamento on lagato notes was identified
as a special case and extracted from the regular legato handling:
these notes are now the only ones to stick to a fixed note pos and
re-use a note object.

During that refactoring, I failed to handle Kit mode properly.
In Kit mode, for each actually used kit item a "slot" in the kitItem
array is reserved and the counter partnote[pos].itemsplaying is incremented,
so the next kit item to handle will use the next "slot" above.

Since legato-portamento notes now re-use an existing note-pos without first
clearing / deleting this pos, we actually need to reset the itemsplaying counter.

WARNING: this whole scheme behaves strangely when the note range of some
kit items is limited. This used to be so in the old code, and the whole
refactoring and also this bugfix here does nothing to address this issue.

# Conflicts:
#	src/version.txt