Small refinements in Main, SynthEngine, MiscGui, LV2
Trying to get a better understanding of LV2 I noticed it used Fl::Check(), whereas in Main we use Fl::Wait(33333). Looking up the
difference I realised I'd been mistaken in how wait operated. It
doesn't wait for the specified time, but will wait NO LONGER than
that time. It's badly named as it's a GUI refresh action, and as
soon as it's made any changes it exits. If nothing has changed, it
exits immediately.

Changing this to using check(), and then always doing usleep(33333)
now gives a fairly stable refresh rate, and a very tiny reduction
in processor load. It also stops the startup VU overload we were
seeing when using autostart multiple instances, so I've been able to remove the checks in SynthEngine().

Looking back in LV2 I cleared out a number of commented out parts so I could see what was happening more clearly. What now puzzles me
is the sleep in idleThread(). We make no use of this, so why does
the function even exist? I've marked it as a TODO.

Putting a few canaries in various places confirmed that the length
of this delay has no effect on the GUI refresh. This is determined
entirely by the host (and they are all slightly different).

I've put a final TODO where there is a Fl::Check() just *after* the
command to close the GUI. It seems behave correctly regardless of
whether it's there or not.