BugFix: kitmode GUI engines not upodating correctly.
Previously the entire part GUI was rebuilt on part changes, whereas
now we update the exisitng structure. However, the individual kit
engines are only updated when their edit button is hit. Also, this is
only done if it's not the same kit number as before - even if it was
previously displaying a different part.

The solution was to set the kit item number to -1 on kit refreshes.