Further work on presets: Fixed clipboard. It was setting XML to full depth without restoring it afterwards.
Removed alternative names. These were intended to be able to fetch
the newer Zyn generated ones, but some of these are now
incompatible.

Simplified presets lists. This was storing the leaf name as well as the full filename. This is trivially easy to find with the filemanager.

PresetsStore is still making direct access to files. Next on the
list to sort out!