Merged in copyPaste so branches don't get out of sync. The changes already made should have no effect on behaviour and the test command is not identifed to users anywhere. Also added more detail the the file manager comments.
Fixed CLI regression - not importing Scala keymaps.
A 'clever' shortcut wasn't so clever!

BugFix: scale imports not reporting names when loaded.
This was trying to read the wrong variable, and in any case it
would have needed two copies of the message.

Modifed TextMsgBuffer.h to provide non-destructive read.
Triggered by the above, but there are other places where this
would be very helpful.