Instances now have independent preset directory choice. Previously these were all slaved to firstSynth.
This required a deep dive into config (in particular) and creation
of a separate file in {usr}.local for the actual list of recognised
preset roots. The existing entries in the main config file are
still there but will not be updated.

Also changes were needed in a number of other significant files
and the opportunity was taken to remove a group of old forward
config transfer code for file locations etc.

This had the side effect of removing about a dozen instances of
firstSynth use and associated extern entries.