BugFix: MIDI-learn correctly saving but not recognising most saved effects controls.
This occurred mainly because the text recognition system was
expecting list entries to have a lead capital, then all lower case
whereas most of these are mixed case. The solution to this was to
make a case insensitive search. Added to this the EQ effect has an
exxeptional numbering sequence that was not catered for.

Behaviour now exactly matches that of the older (still supported)
numeric system, and selection of which to use and whether to report
differences has been improved.

However, there is still a problem. With both forms, the EQ in-band
controls only work if the GUI is displaying that particular band!