BugFix: formant filter window not updating centre frequency and octave range numeric entries.
We have to be careful with GUI callbacks as often they send data, so if they are called from the returns function there
may be a 33mS continuous loop!
In the particular case of the above two, they make no calls
and are not writable from the GUI so the calback must always
be made regardless of the source.

Set version as 2.2.3 rc2