Added two drum sounds from Lorenzo Some Dynfilter cleanups, but more needed later. Everything seems OK apart from the formant filter settings. Values are correct but there is a major problem getting the per-formant controls to reset to the preset's defaults on a right button click or for these controls to indicate their status properly.
I've made several attempts at this, and each time abandoned them as
they were getting vastly over-complicated (and still not right).

The issue is that there is a 'loose' connection between effects and
filters such that it's difficult for the filter to find out what it is
supposed to do over the individual vowels and formants in the case of
VocalMorph1 and VocalMorph2.

The situation is better than it was, but I don't want to spend any more
time on it until after V2.3.0 release - even though I don't like
leaving known issues hanging.
    W.G.