BugFix: Crash in legacy PadSynth 'apply' code. It seems that Pkitmode is not recognised at the point where this is tested (InterChange.cpp line 1535) and switching PadSynth on/off from the main Edit window *always* sets the kitnumber greater than the maximim, so testing this sees to be the simplest way of resoving the issue.
BugFix: Some Windows not recognised as from Yoshimi due to it not being in
the window title. This has been added to all of them, and will be further
improved.

New feature: If the virtual keboard is visible, using the mouse to select
instruments from banks will pass computer key focus to the keyboard so that
different sounds can be compared without having a MIDI keyboard attached.

Updated some docs.