[guiconnect] - Transfer of data into the GUI - setup for prototyping
The goal of this development effort is to remove direct data access by the GUI.
Up to now it is a common pattern for the GUI to reach into internal data of the Core,
which both causes problems with thread safety and hinders a separation of GUI and Core.
While commands and response messages are already dispatched through a command system
based on ringbuffers, there was no suitable way to transfer larger data blocks.

The intended communication pattern is as follows

- some flavour of "publisher-subscriber" in the Core, with PUSH on updates
- a notification message pushed into the GUI through the command ringbuffer
- as part of regular command dispatch, the GUI will also retreave these messages
- ...and then pick up and dispatch the data blocks from a data block buffer

This changeset adds new source files and attaches a test / playground / prototypie
temporarily into the Yoshimi testInvoker; launch it with
cd test
execute