[guiconnect] - decide where to handle the typed context
For this (planned) communication protocol, we want to accept *arbitrary* data blocks.
But we also intend to handle those arbitrary data blocks in a type safe manner.
This implies that each "conversation" needs to run within a specifically typed context,
while the communication mechanism as such remains type agnostic.

However, the actual problem is how to /transport/ this typed context from the
producer to the consumer(s). An elegant (yet slightly ambitious) soulution would be
to transport this information /statically/, i.e. at compilation time. This
can be achieved by turning the Connection handle into a template; both ends
of the actual connection must know about the actual data anyway, so they can
use a specific instance of this Connection template marked with this actual
data type. To allow for safety checks, the type info will be translated into
an opaque marker (hash), to be verified at the receiving end before accessing
the transported data