[guiconnect] - refactoring to prepare for synchronisation
This is a preparation for *any kind* of thread visibility synchronisation.
Irrespective, if the actual sync is done with a mutex, or thorugh atomics,
in any case we must ensure that the complete sequence to allocate the next
storage slot can be enclosed into a "synchronisation bracket".

This turns out somewhat tricky, because the actual invocation of
GuiDataExchange::Connection::publish(DAT) is "templated", which means
the compiler will generate a distinct instance of that function for each
different type used with GuiDataExchange. On the other hand, we have a common
"implementation back end" in GuiDataExchange.cpp, which is type agnostic.

The solution is to package the actual copying of data into a lambda-function,
and pass this wrapped into a std::function<void(void*)> down into the backend.
This allows then to confine the technicalities of data synchronisation to the
backend implementation.

I have verified (by looking at the assembly using Godbolt.org) that current optimisers
are able to remove most of the indirection overhead and in fact neither construct the
std::function object at all, nor perform any spurious heap allocations. At runtime,
only a simple function pointer to the generated copy-λ for each type is passed down.

Moreover, with this changeset, a bit of template-metaprogramming is introduced
to derive the required buffer size automatically at compile time. For this to work,
all relevant data types, which shall be published through GuiDataExchange,
must be included into the definition of the constant "SIZ" in GuiDataExchange.cpp